refactor: move app components to /components/app
