+++ src/app/app/calendar/create/page.tsx
@@ -1,4 +1,4 @@
-import CalendarCreatePage from "@/components/calendar/create-page"
+import CalendarCreatePage from "@/components/app/calendar/create-page"
 
 export default function Page() {
   return <CalendarCreatePage />
