+++ src/app/app/calendar/page.tsx
@@ -1,4 +1,4 @@
-import CalendarOverviewPage from "@/components/calendar/overview-page"
+import CalendarOverviewPage from "@/components/app/calendar/overview-page"
 
 export default function Page() {
   return <CalendarOverviewPage />
