+++ src/app/app/customer/page.tsx
@@ -1,4 +1,4 @@
-import CustomerOverviewPage from "@/components/customer/overview-page"
+import CustomerOverviewPage from "@/components/app/customer/overview-page"
 
 export default function Page() {
   return <CustomerOverviewPage />
