+++ src/app/app/employee/create/page.tsx
@@ -1,4 +1,4 @@
-import EmployeeCreatePage from "@/components/employee/create-page"
+import EmployeeCreatePage from "@/components/app/employee/create-page"
 
 export default function Page() {
   return <EmployeeCreatePage />
