+++ src/app/app/employee/page.tsx
@@ -1,4 +1,4 @@
-import EmployeeOverviewPage from "@/components/employee/overview-page"
+import EmployeeOverviewPage from "@/components/app/employee/overview-page"
 
 export default function Page() {
   return <EmployeeOverviewPage />
