+++ src/app/app/job/create/page.tsx
@@ -1,4 +1,4 @@
-import JobCreatePage from "@/components/job/create-page"
+import JobCreatePage from "@/components/app/job/create-page"
 
 export default function Page() {
   return <JobCreatePage />
