+++ src/app/app/job/page.tsx
@@ -1,4 +1,4 @@
-import JobOverviewPage from "@/components/job/overview-page"
+import JobOverviewPage from "@/components/app/job/overview-page"
 
 export default function Page() {
   return <JobOverviewPage />
