+++ src/app/auth/login/page.tsx
@@ -1,4 +1,4 @@
-import { LoginPage } from "@/components/auth/login-page"
+import { LoginPage } from "@/components/app/auth/login-page"
 
 export default function Page() {
   return <LoginPage />
