+++ src/components/app/customer/details-form.tsx
